+++ packages/gatsby-plugin-orion-core/src/components/Form/index.stories.js
@@ -1,7 +1,8 @@
 import React from 'react'
 import { storiesOf } from '@storybook/react'
 import { jsxDecorator } from 'storybook-addon-jsx'
-import { Typography } from '@material-ui/core'
+import { object } from '@storybook/addon-knobs'
+import { Typography, Button } from '@material-ui/core'
 
 import Form from '.'
 
@@ -12,29 +13,28 @@ storiesOf('Form', module)
       formFields={[
         {
           label: 'First Name',
-          key: 'given_name',
+          name: 'given_name',
           required: true,
           type: 'text',
           xs: 6,
         },
         {
           label: 'Last Name',
-          key: 'family_name',
+          name: 'family_name',
           required: true,
           type: 'text',
           xs: 6,
         },
         {
           label: 'Email',
-          key: 'username',
+          name: 'username',
           required: true,
           type: 'email',
           xs: 12,
-          error: 'This is an error',
         },
         {
           label: 'Password',
-          key: 'password',
+          name: 'password',
           required: true,
           type: 'password',
           xs: 12,
@@ -42,13 +42,7 @@ storiesOf('Form', module)
             'Must include an uppercase character, a number and a symbol',
         },
       ]}
-      Title={<Typography variant="h3">Create a new account</Typography>}
-      submitButtonText="Create account"
-      submitText={
-        <Typography noWrap variant="h6" color="textSecondary">
-          Have an account? <a>Sign in</a>
-        </Typography>
-      }
+      title={<Typography variant="h3">Create a new account</Typography>}
     />
   ))
   .add('Login Error', () => (
@@ -56,25 +50,19 @@ storiesOf('Form', module)
       formFields={[
         {
           label: 'Please enter your username',
-          key: 'username',
+          name: 'username',
           type: 'email',
           xs: 12,
           error: 'This is an error!',
         },
         {
           label: 'Please enter your password',
-          key: 'password',
+          name: 'password',
           type: 'password',
           xs: 12,
         },
       ]}
-      Title={<Typography variant="h3">Sign in to your account</Typography>}
-      submitButtonText="Sign in"
-      submitText={
-        <Typography noWrap variant="h6" color="textSecondary">
-          No account? <a>Create account</a>
-        </Typography>
-      }
+      title={<Typography variant="h3">Sign in to your account</Typography>}
     />
   ))
   .add('Login', () => (
@@ -82,18 +70,18 @@ storiesOf('Form', module)
       formFields={[
         {
           label: 'Please enter your username',
-          key: 'username',
+          name: 'username',
           type: 'email',
           xs: 12,
         },
         {
           label: 'Please enter your password',
-          key: 'password',
+          name: 'password',
           type: 'password',
           xs: 12,
         },
       ]}
-      Title={<Typography variant="h3">Sign in to your account</Typography>}
+      title={<Typography variant="h3">Sign in to your account</Typography>}
       submitButtonText="Sign in"
       submitText={
         <Typography noWrap variant="h6" color="textSecondary">
@@ -102,5 +90,65 @@ storiesOf('Form', module)
       }
       // eslint-disable-next-line no-console
       onSubmit={e => console.log(e)}
+    />
+  ))
+  .add('Article Edit', () => (
+    <Form
+      formFields={object('Form Fields', [
+        {
+          label: 'Please enter your username',
+          name: 'username',
+          multiline: true,
+          required: true,
+          rows: 10,
+          inputTypographyVariant: 'h1',
+          xs: 6,
+          placeholder: 'addTitle',
+        },
+        {
+          label: 'Please enter your password',
+          name: 'password',
+          type: 'password',
+          validate: v => (v.length < 5 ? 'Password too short' : null),
+          xs: 6,
+        },
+        {
+          label: 'Please enter your username',
+          name: 'email',
+          type: 'email',
+          inputTypographyVariant: 'h2',
+          placeholder: 'Email',
+          validate: value => {
+            let error
+
+            if (!value) {
+              error = 'Required'
+            } else if (
+              !/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i.test(value)
+            ) {
+              error = 'Invalid email address'
+            }
+
+            return error
+          },
+          xs: 12,
+        },
+        {
+          label: 'Please enter your password',
+          name: 'confirmPassword',
+          type: 'password',
+          validate: (value, values) =>
+            value === values.password ? null : 'Passwords do not match',
+          xs: 12,
+        },
+      ])}
+      title={<Typography variant="h3">Sign in to your account</Typography>}
+      SubmitComponent={props => (
+        <Button type="submit" variant="contained" color="primary" {...props}>
+          Custom Submit
+        </Button>
+      )}
+      // eslint-disable-next-line no-alert
+      onSubmit={e => alert(JSON.stringify(e, null, 2))}
     />
   ))