+++ packages/gatsby-plugin-orion-core/src/components/SearchInput/index.stories.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import { storiesOf } from '@storybook/react'
+import { jsxDecorator } from 'storybook-addon-jsx'
+import SearchInput from '.'
+
+storiesOf('SearchInput', module)
+  .addDecorator(jsxDecorator)
+  .add('Default', () => (
+    <div style={{ width: 256 }}>
+      <SearchInput onSearch={() => {}} />
+    </div>
+  ))